Fixed issue with empty lists matching infinite lists
The problem was, that the previous implementation always tried to match
something as long as the result was false and the lhs expression
changed. Infinite lists always change, therefor this case had to be
handled differently: Now a new return type 'None' is returned, if the
match is definitly false.